+++ packages/cli/src/cli.ts
@@ -145,6 +145,10 @@ export async function cli() {
             "-prr, --pull-request-reviews",
             "create pull request reviews from annotations"
         )
+        .option(
+            "-tm, --teams-message",
+            "Posts a message to the teams channel"
+        )
         .option("-j, --json", "emit full JSON response to output")
         .option("-y, --yaml", "emit full YAML response to output")
         .option(`-fe, --fail-on-errors`, `fails on detected annotation error`)
