+++ packages/cli/src/runtime.ts
@@ -310,152 +310,3 @@ export async function markdownifyPdf(
 
     return { pages, images, markdowns }
 }
-
-function parseTeamsChannelUrl(url: string) {
-    const m =
-        /^https:\/\/teams.microsoft.com\/[^\/]{1,32}\/channel\/(?<channelId>.+)\/.*\?groupId=(?<teamId>([a-z0-9\-])+)$/.exec(
-            url
-        )
-    if (!m) throw new Error("Invalid Teams channel URL")
-    const { teamId, channelId } = m.groups
-    return { teamId, channelId }
-}
-
-let _azureToken: string
-async function azureGetToken(): Promise<string> {
-    if (!_azureToken) {
-        console.debug(`Azure: retreiving token...`)
-        const { DefaultAzureCredential } = await import("@azure/identity")
-        const credential = new DefaultAzureCredential()
-        const tokenResponse = await credential.getToken(
-            "https://graph.microsoft.com/.default"
-        )
-        if (!tokenResponse) throw new Error("Failed to retrieve access token.")
-        _azureToken = tokenResponse.token
-    }
-    return _azureToken
-}
-
-export interface MicrosoftTeamsEntity {
-    webUrl: string
-    name: string
-}
-
-/**
- * Uploads a file to the files storage of a Microsoft Teams channel.
- * @param channelUrl Shared channel link in the format https://teams.microsoft.com/l/channel/<channelId>/<channelName>?groupId=<teamId>
- * @param filename
- * @returns
- */
-async function microsoftTeamsChannelUploadFile(
-    token: string,
-    channelUrl: string,
-    filename: string
-): Promise<MicrosoftTeamsEntity> {
-    console.debug(`Uploading ${filename}...`)
-
-    const { teamId, channelId } = parseTeamsChannelUrl(channelUrl)
-    const Authorization = `Bearer ${token}`
-
-    const channelInfoUrl = `https://graph.microsoft.com/v1.0/teams/${teamId}/channels/${channelId}`
-    const channelInfoRes = await fetch(channelInfoUrl, {
-        headers: {
-            Authorization,
-        },
-    })
-    if (!channelInfoRes.ok) {
-        throw new Error(
-            `Failed to get channel info: ${channelInfoRes.status} ${channelInfoRes.statusText}`
-        )
-    }
-    const channelInfo = await channelInfoRes.json()
-    const folder = channelInfo.displayName
-
-    // resolve channel folder name
-    const file = await readFile(filename)
-    const url = `https://graph.microsoft.com/v1.0/groups/${teamId}/drive/root:/${folder}/${path.basename(
-        filename
-    )}:/content`
-    const res = await fetch(url, {
-        method: "PUT",
-        headers: {
-            Authorization,
-            "Content-Type": "application/octet-stream",
-        },
-        body: file,
-    })
-    if (!res.ok) {
-        console.debug(await res.text())
-        throw new Error(
-            `Failed to upload file: ${res.status} ${res.statusText}`
-        )
-    }
-    const j = (await res.json()) as MicrosoftTeamsEntity
-    return j
-}
-
-export async function microsoftTeamsChannelPostMessage(
-    channelUrl: string,
-    subject: string,
-    message: string,
-    options?: {
-        files?: string[]
-    }
-): Promise<MicrosoftTeamsEntity> {
-    const { files = [] } = options || {}
-    const { teamId, channelId } = parseTeamsChannelUrl(channelUrl)
-    const token = await azureGetToken()
-
-    const body = {
-        body: {
-            contentType: "html",
-            content: message,
-        },
-        subject,
-        attachments: [] as any[],
-    }
-
-    for (const file of files) {
-        const fres = await microsoftTeamsChannelUploadFile(
-            token,
-            channelUrl,
-            file
-        )
-        const guid = crypto.randomUUID()
-        body.body.content += "\n" + `<attachment id=\"${guid}\"></attachment>`
-        body.attachments = [
-            {
-                id: guid,
-                contentType: "reference",
-                contentUrl: fres.webUrl,
-                name: fres.name,
-                thumbnailUrl: null,
-            },
-        ]
-    }
-
-    const url = `https://graph.microsoft.com/v1.0/teams/${teamId}/channels/${channelId}/messages`
-    const response = await fetch(url, {
-        method: "POST",
-        headers: {
-            Authorization: `Bearer ${token}`,
-            "Content-Type": "application/json",
-        },
-        body: JSON.stringify(body),
-    })
-
-    if (!response.ok) {
-        const err: any = await response.text()
-        console.debug(err)
-        throw new Error(
-            `Failed to post message: ${response.status} ${response.statusText}`
-        )
-    }
-
-    {
-        const data: any = await response.json()
-        const { webUrl } = data
-        console.debug(`message created at ${webUrl}`)
-        return data
-    }
-}
