+++ packages/core/src/testhost.ts
@@ -64,6 +64,8 @@ export class TestHost implements RuntimeHost {
     // File system for workspace
     workspace: WorkspaceFileSystem
     azureToken: AzureTokenResolver = undefined
+    azureServerlessToken: AzureTokenResolver = undefined
+    microsoftGraphToken: AzureTokenResolver = undefined
 
     // Default options for language models
     readonly modelAliases: ModelConfigurations = defaultModelConfigurations()
